+++ src/cachelib/file.py
@@ -1,4 +1,5 @@
 import errno
+import hashlib
 import logging
 import os
 import platform
@@ -7,13 +8,20 @@ import struct
 import tempfile
 import typing as _t
 from contextlib import contextmanager
-from hashlib import md5
 from pathlib import Path
 from time import sleep
 from time import time
 
 from cachelib.base import BaseCache
 from cachelib.serializers import FileSystemSerializer
+
+
+def _lazy_md5(string: bytes = b"") -> _t.Any:
+    """Don't access ``hashlib.md5`` until runtime. FIPS builds may not include
+    md5, in which case the import and use as a default would fail before the
+    developer can configure something else.
+    """
+    return hashlib.md5(string)
 
 
 class FileSystemCache(BaseCache):
@@ -32,12 +40,16 @@ class FileSystemCache(BaseCache):
     :param mode: the file mode wanted for the cache files, default 0600
     :param hash_method: Default hashlib.md5. The hash method used to
                         generate the filename for cached results.
+                        Default is lazy loaded and can be overridden by
+                        setting  `_default_hash_method`
     """
 
     #: used for temporary files by the FileSystemCache
     _fs_transaction_suffix = ".__wz_cache"
     #: keep amount of files in a cache element
     _fs_count_file = "__wz_cache_count"
+    #: default file name hashing method
+    _default_hash_method = staticmethod(_lazy_md5)
 
     serializer = FileSystemSerializer()
 
@@ -47,12 +59,15 @@ class FileSystemCache(BaseCache):
         threshold: int = 500,
         default_timeout: int = 300,
         mode: _t.Optional[int] = None,
-        hash_method: _t.Any = md5,
+        hash_method: _t.Any = None,
     ):
         BaseCache.__init__(self, default_timeout)
         self._path = cache_dir
         self._threshold = threshold
-        self._hash_method = hash_method
+
+        self._hash_method = self._default_hash_method
+        if hash_method is not None:
+            self._hash_method = hash_method
 
         # Mode set by user takes precedence. If no mode has
         # been given, we need to set the correct default based
@@ -148,10 +163,7 @@ class FileSystemCache(BaseCache):
                     exc_info=True,
                 )
         fname_sorted = (
-            fname
-            for _, fname in sorted(
-                exp_fname_tuples, key=lambda item: item[0]  # type: ignore
-            )
+            fname for _, fname in sorted(exp_fname_tuples, key=lambda item: item[0])
         )
         for fname in fname_sorted:
             try:
@@ -302,7 +314,9 @@ class FileSystemCache(BaseCache):
             )
             return False
 
-    def _run_safely(self, fn: _t.Callable, *args: _t.Any, **kwargs: _t.Any) -> _t.Any:
+    def _run_safely(
+        self, fn: _t.Callable[..., _t.Any], *args: _t.Any, **kwargs: _t.Any
+    ) -> _t.Any:
         """On Windows os.replace, os.chmod and open can yield
         permission errors if executed by two different processes."""
         if platform.system() == "Windows":
@@ -326,7 +340,9 @@ class FileSystemCache(BaseCache):
         return output
 
     @contextmanager
-    def _safe_stream_open(self, path: str, mode: str) -> _t.Generator:
+    def _safe_stream_open(
+        self, path: str, mode: str
+    ) -> _t.Generator[_t.Any, None, None]:
         fs = self._run_safely(open, path, mode)
         if fs is None:
             raise OSError