+++ tests/test_interface_uniformity.py
@@ -12,7 +12,7 @@ from cachelib import SimpleCache
 
 @pytest.fixture(autouse=True)
 def create_cache_list(request, tmpdir):
-    mc = MemcachedCache()
+    mc = MemcachedCache(servers=["127.0.0.1:11212"])
     if mc.pylibmc_used:
         with mc._client.reserve(block=mc.blocking) as client:
             client.flush_all()