+++ CHANGES.rst
@@ -1,3 +1,40 @@
+Version 0.14.0
+--------------
+
+Unreleased
+
+- improve typing and refresh project to use pyproject and pallets workflows :pr:`420`
+- replace deprecated ``datetime.utcnow()`` with ``datetime.now()`` :pr:`421`
+
+Version 0.13.0
+--------------
+
+Released 2024-04-13
+
+-   default ``hashlib.md5`` may not be available in FIPS builds. We
+    now do not access it at import time on ``FileSystemCache``so developers
+    have time to change the default.
+    ``hashlib.md5`` will be lazy loaded when a new default is not provided
+
+
+Version 0.12.0
+--------------
+
+Released 2024-02-11
+
+- ``RedisCache`` now supports callables as keys
+- Added ``MongoDB`` as a cache backend
+
+
+Version 0.11.0
+--------------
+
+Released 2024-02-10
+
+- Drop python 3.7 support
+- Add python 3.11 support
+
+
 Version 0.10.2
 --------------
 
@@ -9,7 +46,7 @@ Released 2023-01-31
 Version 0.10.1
 --------------
 
-Unreleased
+Released 2023-01-22
 
 - Fix logging pollution due to ``DynamoDB`` logging handler
 
@@ -21,7 +58,6 @@ Released 2023-01-22
 
 - Improve error message when ``FileSystemCache`` methods are called with non-str keys. :pr:`170`
 - Added ``DynamoDb`` as a cache backend :pr:`209`
-
 
 Version 0.9.0
 -------------
@@ -104,7 +140,7 @@ Released 2021-08-12
     lost. :pr:`52`
 -   ``FileSystemCache`` logs system-related exceptions. :pr:`51`
 -   Removal of expired entries in ``FileSystemCache`` is only triggered
-    if the number of entries is over the ``threshhold`` when calling
+    if the number of entries is over the ``threshold`` when calling
     ``set``. ``get`` ``has`` still return ``None`` and ``False``
     respectively for expired entries, but will not remove the files. All
     removals happen at pruning time or explicitly with ``clear`` and